+++ src/main/java/beyondvision/record/dto/request/ExerciseRecordRequest.java
@@ -13,7 +13,7 @@ public class ExerciseRecordRequest {
     private Integer exerciseTime;
     private Integer exerciseCount;
     private Integer successCount;
-    private Integer caloriesBurnedSum;
+    private Double caloriesBurnedSum;
     private Integer averageHeartRate;
 
     @NotNull(message = "회원 id는 필수입니다.")
